requirements.txt
Django>=5.0,<6.0
Pillow>=10.0.0
django-crispy-forms>=2.1
crispy-bootstrap5>=2024.2
